Created actual UI, removed unneeded features
Added UI that can be used to load / reload emulator. Options are now generated in appdata and all assets are packaged with the .exe, so it can be placed and ran anywhere in the filesystem with no dependencies. Also removed unneeded logging and instruction lookup, because who cares.

TO-DO: options menu, fix saving / loading